Move hash-to-curve tests to `tests/` for compilation performance
Extracted from zkcrypto/bls12_381#90.
